+++ src/main/java/dev/willyelton/crystal_tools/utils/ItemStackUtils.java
@@ -2,6 +2,7 @@ package dev.willyelton.crystal_tools.utils;
 
 import dev.willyelton.crystal_tools.gui.ModGUIs;
 import dev.willyelton.crystal_tools.gui.UpgradeScreen;
+import dev.willyelton.crystal_tools.tool.LevelableItem;
 import dev.willyelton.crystal_tools.tool.LevelableTool;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
@@ -9,7 +10,7 @@ import net.minecraft.world.item.ItemStack;
 public class ItemStackUtils {
     public static ItemStack getHeldLevelableTool(Player player) {
         for (ItemStack i : player.getHandSlots()) {
-            if (i.getItem() instanceof LevelableTool) {
+            if (i.getItem() instanceof LevelableItem) {
                 return i;
             }
         }