+++ changelog.md
@@ -1,3 +1,18 @@
+----------------------- 0.4.0 Changes -----------------------
+### Additions
+- Added the crystal Bow
+- Added the following attributes for it
+  - Arrow Damage
+  - Arrow Speed
+  - Flame
+  - Infinity
+  - Arrow Knockback
+- I planned to add draw speed as well, but that is not working right now, will be added later
+- Texture to come soon
+
+### Changes
+- Updated to Forge 1.18.1-39.0.59
+
 ----------------------- 0.3.0 Changes -----------------------
 ### Additions
 - Added the crystal Sword